Warnings

+++ gview.cpp
@@ -23,7 +23,7 @@ bool gViewLog=false;
 void not_used(void *p)
 {
     if (gViewLog)
-        qDebug() << "Not used ptr "<<(int)p;
+        qDebug() << "Not used ptr " << p;
 
     //this message must never apear if no errors
 }
